Refatorando classes e ajeitando relacionamentos

+++ database/migrations/2024_11_15_010106_remove_user_id_from_users_table.php
@@ -9,10 +9,12 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-    public function up(): void
+    public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('user_id');
+            if (Schema::hasColumn('users', 'user_id')) { // Verifica se a coluna existe antes de remover
+                $table->dropColumn('user_id');
+            }
         });
     }
